pyproject.toml
[project]
name = "rag-v2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.11.0",
    "altair>=5.5.0",
    "arize-phoenix>=12.9.0",
    "arize-phoenix-client>=1.21.0",
    "authlib>=1.6.5",
    "beautifulsoup4>=4.14.2",
    "django>=5.2.7",
    "docker>=7.1.0",
    "docling>=2.60.0",
    "docling-core>=2.50.0",
    "docling-ibm-models>=3.10.2",
    "docling-parse>=4.7.0",
    "docx2txt>=0.9",
    "extra-streamlit-components>=0.1.81",
    "faiss-cpu>=1.12.0",
    "faker>=37.12.0",
    "fastapi>=0.120.4",
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-docling>=1.1.0",
    "langchain-experimental==0.3.4",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langsmith>=0.4.39",
    "lxml>=6.0.2",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "openai>=2.6.1",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "pillow>=11.3.0",
    "psutil>=7.1.2",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "pydeck>=0.9.1",
    "pypdf>=6.1.3",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "rapidocr>=3.4.2",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.50.0",
    "streamlit-authenticator>=0.4.2",
    "sympy>=1.14.0",
    "tabulate>=0.9.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "typer>=0.19.2",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
    "xlsxwriter>=3.2.9",
]
